moved SequenceK to arrow-core-data
