+++ modules/core/arrow-extras-data/src/test/kotlin/arrow/data/SequenceKTest.kt
@@ -1,17 +1,22 @@
 package arrow.data
 
 import arrow.Kind
+import arrow.core.SequenceK
 import arrow.core.Tuple2
+import arrow.core.ForSequenceK
 import arrow.core.extensions.eq
 import arrow.core.extensions.hash
-import arrow.data.extensions.sequencek.applicative.applicative
-import arrow.data.extensions.sequencek.eq.eq
-import arrow.data.extensions.sequencek.hash.hash
-import arrow.data.extensions.sequencek.monad.monad
-import arrow.data.extensions.sequencek.monoid.monoid
-import arrow.data.extensions.sequencek.monoidK.monoidK
-import arrow.data.extensions.sequencek.monoidal.monoidal
-import arrow.data.extensions.sequencek.traverse.traverse
+import arrow.core.k
+import arrow.core.toList
+import arrow.core.fix
+import arrow.core.extensions.sequencek.applicative.applicative
+import arrow.core.extensions.sequencek.eq.eq
+import arrow.core.extensions.sequencek.hash.hash
+import arrow.core.extensions.sequencek.monad.monad
+import arrow.core.extensions.sequencek.monoid.monoid
+import arrow.core.extensions.sequencek.monoidK.monoidK
+import arrow.core.extensions.sequencek.monoidal.monoidal
+import arrow.core.extensions.sequencek.traverse.traverse
 import arrow.test.UnitSpec
 import arrow.test.generators.sequenceK
 import arrow.test.laws.HashLaws
