+++ modules/core/arrow-mtl/src/main/kotlin/arrow/mtl/extensions/sequencek.kt
@@ -4,15 +4,15 @@ import arrow.Kind
 import arrow.core.Either
 import arrow.core.Option
 import arrow.core.Tuple2
-import arrow.data.ForSequenceK
-import arrow.data.SequenceK
-import arrow.data.SequenceKOf
-import arrow.data.fix
+import arrow.core.ForSequenceK
+import arrow.core.SequenceK
+import arrow.core.SequenceKOf
+import arrow.core.fix
 import arrow.extension
 import arrow.mtl.typeclasses.FunctorFilter
 import arrow.mtl.typeclasses.MonadCombine
 import arrow.mtl.typeclasses.MonadFilter
-import arrow.data.combineK as sequenceCombineK
+import arrow.core.combineK as sequenceCombineK
 
 @extension
 interface SequenceKMonadCombine : MonadCombine<ForSequenceK> {
