+++ modules/core/arrow-test/src/main/kotlin/arrow/test/generators/Generators.kt
@@ -28,7 +28,7 @@ import arrow.core.extensions.option.functor.functor
 import arrow.core.ListK
 import arrow.core.MapK
 import arrow.core.NonEmptyList
-import arrow.data.SequenceK
+import arrow.core.SequenceK
 import arrow.core.SetK
 import arrow.data.SortedMapK
 import arrow.core.Validated
