+++ modules/optics/arrow-optics/src/main/kotlin/arrow/optics/extensions/sequencek.kt
@@ -4,8 +4,8 @@ import arrow.Kind
 import arrow.core.left
 import arrow.core.right
 import arrow.core.toT
-import arrow.data.SequenceK
-import arrow.data.k
+import arrow.core.SequenceK
+import arrow.core.k
 import arrow.extension
 import arrow.optics.Optional
 import arrow.optics.POptional
