+++ modules/optics/arrow-optics/src/test/kotlin/arrow/optics/instances/SequenceKInstanceTest.kt
@@ -3,10 +3,10 @@ package arrow.optics.instances
 import arrow.core.Option
 import arrow.core.extensions.eq
 import arrow.core.ListK
-import arrow.data.SequenceK
+import arrow.core.SequenceK
 import arrow.core.extensions.listk.eq.eq
 import arrow.core.extensions.option.eq.eq
-import arrow.data.extensions.sequencek.eq.eq
+import arrow.core.extensions.sequencek.eq.eq
 import arrow.optics.extensions.sequencek.each.each
 import arrow.optics.extensions.sequencek.filterIndex.filterIndex
 import arrow.optics.extensions.sequencek.index.index
